# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gratification API",
        "description": "Documentação da API do Gratification",
        "contact": {
            "email": "suporte@gratification.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/login": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Display the login view",
                "description": "Display the login view.",
                "operationId": "d38b0625602ec8e41c3bb98aa2e969de",
                "responses": {
                    "200": {
                        "description": "Login page displayed"
                    }
                }
            },
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate a user",
                "description": "Handle an incoming authentication request.",
                "operationId": "42f77e072dec84b0e1094cca1228298e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User authenticated successfully"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
        "/email/verification-notification": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Send a new email verification notification",
                "description": "Send a new email verification notification.",
                "operationId": "7c01a354a5960445207bc5d2f17d547e",
                "responses": {
                    "200": {
                        "description": "Email verification link sent"
                    },
                    "400": {
                        "description": "Email already verified or error sending verification link"
                    }
                }
            }
        },
        "/email/verify": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Display the email verification prompt",
                "description": "Display the email verification prompt.",
                "operationId": "4b827d276f46c16e5fc96596a67d53cd",
                "responses": {
                    "200": {
                        "description": "Email verification prompt displayed"
                    },
                    "302": {
                        "description": "Redirect to dashboard if email is already verified"
                    }
                }
            }
        },
        "/reset-password/{token}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Display the password reset view",
                "description": "Display the password reset view.",
                "operationId": "139cbd7a33506f23db79a42dd3a1b825",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Password reset token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password reset view displayed"
                    }
                }
            }
        },
        "/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Handle password reset request",
                "description": "Handle an incoming new password request.",
                "operationId": "24b1c68468bf984b8806908f211f3827",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "reset-token"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset successfully"
                    },
                    "400": {
                        "description": "Error in password reset"
                    }
                }
            }
        },
        "/user/password": {
            "put": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Update the user's password",
                "description": "Update the user's password.",
                "operationId": "b2eb628b6e49b9a32146aae16edcf3f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "oldpassword123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newpassword123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password updated successfully"
                    },
                    "400": {
                        "description": "Invalid current password or invalid data"
                    }
                }
            }
        },
...
        "/email/verify/{id}/{hash}": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify the user's email address",
                "description": "Mark the authenticated user's email address as verified.",
                "operationId": "defff15cf49d15083f3d7d2284955a29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "Verification hash",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verified successfully"
                    },
                    "400": {
                        "description": "Error during email verification"
                    }
                }
            }
        },
...
    },
...
}